+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2026_06_02_000100) do
+ActiveRecord::Schema[8.1].define(version: 2026_06_03_000100) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_catalog.plpgsql"
 
@@ -24,12 +24,14 @@ ActiveRecord::Schema[8.1].define(version: 2026_06_02_000100) do
     t.datetime "last_imported_at"
     t.datetime "log_cursor_at"
     t.string "name", null: false
+    t.boolean "skip_analysis", default: false, null: false
     t.string "state"
     t.datetime "updated_at", null: false
     t.index ["active"], name: "index_docker_containers_on_active"
     t.index ["docker_id"], name: "index_docker_containers_on_docker_id", unique: true
     t.index ["import_status"], name: "index_docker_containers_on_import_status"
     t.index ["name"], name: "index_docker_containers_on_name"
+    t.index ["skip_analysis"], name: "index_docker_containers_on_skip_analysis"
   end
 
   create_table "inference_settings", force: :cascade do |t|