+++ test/jobs/analyze_log_entry_job_test.rb
@@ -19,4 +19,18 @@ class AnalyzeLogEntryJobTest < ActiveJob::TestCase
       AnalyzeLogEntryJob.perform_now(-1)
     end
   end
+
+  test 'skips analysis for excluded containers' do
+    entry = log_entries(:pending_warning)
+    docker_containers(:web).update!(name: entry.source_container, skip_analysis: true)
+    called = false
+    original_call = LogEntries::Analyzer.method(:call)
+    LogEntries::Analyzer.define_singleton_method(:call) { |_entry| called = true }
+
+    AnalyzeLogEntryJob.perform_now(entry.id)
+
+    assert_not called
+  ensure
+    LogEntries::Analyzer.define_singleton_method(:call, original_call)
+  end
 end